refactor problem to GeometryBasics.Mesh conversion
