+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 
-## 0.0.1 (2021-01-xx)
+## Unreleased (2021-02-xx)
 
 ### Added
-* Added `buckling` functions to `TopOptProblems`
+* Added `buckling` functions to `TopOptProblems`, but not well tested yet.
+* Added `visualization` method for visualization FEA deformation results.
 
 ### Changed
-*
+* Wrapped all visualization methods into the `TopOptProblems.Visualization` module
 
 ### Fixed
 *
+
+## 0.1.0 (2021-02-08)
+
+Initial Release.
